requirements.txt
Flask>=1.0.0
google-cloud-datastore==1.7.1
google-cloud-storage==1.13.0
google-cloud-logging==1.8.0
google-cloud-error_reporting==0.30.0
gunicorn==19.9.0
oauth2client==4.1.3
Flask-SQLAlchemy==2.3.2
PyMySQL==0.9.2
Flask-PyMongo>=2.0.0
PyMongo==3.7.2
six==1.11.0
requests[security]==2.20.1
honcho==1.0.1
psq==0.7.0
python-dotenv
